fixed crash/misbehaviour in cross breed dialogue * open a library and select a patch * select Cross Breed from the menu * select Keep to copy whatever you generated into the library * select Generate to make a new tone. nothing happens because of a trapped crash
when the generated patch is pasted into the library via Keep, it notifies that the entire table has changed. this causes the selection to be reset, which causes Generate to fail as it doesn't successfully find the initial patch to breed.

I have modified the table notification so that it will notify that a patch has been added (instead of saying the whole table has changed), which prevents the selection being reset.